backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
boto3>=1.34.0
runpod>=1.7.0
python-dotenv>=1.0.0
